backend/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
pyjwt
passlib[bcrypt]
python-multipart
scikit-learn
numpy
pydantic==2.5.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
groq
